AGT-389: Refactoring storageUtils
