+++ libraries/intentIqUtils/storageUtils.js
@@ -1,8 +1,12 @@
 import {logError, logInfo} from '../../src/utils.js';
 import {SUPPORTED_TYPES, FIRST_PARTY_KEY} from '../../libraries/intentIqConstants/intentIqConstants.js';
+import {getStorageManager} from '../../src/storageManager.js';
+import {MODULE_TYPE_UID} from '../../src/activities/modules.js';
 
 const MODULE_NAME = 'intentIqId';
 const PCID_EXPIRY = 365;
+
+export const storage = getStorageManager({moduleType: MODULE_TYPE_UID, moduleName: MODULE_NAME});
 
 /**
  * Read data from local storage or cookie based on allowed storage types.
@@ -10,7 +14,7 @@ const PCID_EXPIRY = 365;
  * @param {Array} allowedStorage - Array of allowed storage types ('html5' or 'cookie').
  * @return {string|null} - The retrieved data or null if an error occurs.
  */
-export function readData(key, allowedStorage, storage) {
+export function readData(key, allowedStorage) {
   try {
     if (storage.hasLocalStorage() && allowedStorage.includes('html5')) {
       return storage.getDataFromLocalStorage(key);
@@ -30,10 +34,9 @@ export function readData(key, allowedStorage, storage) {
  * @param {string} key - The key under which the data will be stored.
  * @param {string} value - The value to be stored (e.g., IntentIQ ID).
  * @param {Array} allowedStorage - An array of allowed storage types: 'html5' for Local Storage and/or 'cookie' for Cookies.
- * @param {Object} storage - The storage handler object that manages Local Storage and Cookies.
  * @param {Object} firstPartyData - Contains user consent data; if isOptedOut is true, data will not be stored (except for FIRST_PARTY_KEY).
  */
-export function storeData(key, value, allowedStorage, storage, firstPartyData) {
+export function storeData(key, value, allowedStorage, firstPartyData) {
   try {
     if (firstPartyData?.isOptedOut && key !== FIRST_PARTY_KEY) {
       return;
@@ -47,6 +50,25 @@ export function storeData(key, value, allowedStorage, storage, firstPartyData) {
         const expiresStr = (new Date(Date.now() + (PCID_EXPIRY * (60 * 60 * 24 * 1000)))).toUTCString();
         storage.setCookie(key, value, expiresStr, 'LAX');
       }
+    }
+  } catch (error) {
+    logError(error);
+  }
+}
+
+/**
+ * Remove Intent IQ data from cookie or local storage
+ * @param key
+ */
+
+export function removeDataByKey(key, allowedStorage) {
+  try {
+    if (storage.hasLocalStorage() && allowedStorage.includes('html5')) {
+      storage.removeDataFromLocalStorage(key);
+    }
+    if (storage.cookiesAreEnabled() && allowedStorage.includes('cookie')) {
+      const expiredDate = new Date(0).toUTCString();
+      storage.setCookie(key, '', expiredDate, 'LAX');
     }
   } catch (error) {
     logError(error);
